split transderToLX to files
