+++ src/components/Features/Transfer/Transfer.js
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
 
 import {ActionType, NetworkType} from '../../../enums';
-import {useMaxAmount, useTransferToL1, useTransferToL2} from '../../../hooks';
+import {useMaxDeposit, useTransferToL1, useTransferToL2} from '../../../hooks';
 import {useL1Token, useL2Token, useTokens} from '../../../providers/TokensProvider';
 import {
   Loading,
@@ -33,7 +33,7 @@ export const Transfer = () => {
   const transferToL1 = useTransferToL1();
   const getL1Token = useL1Token();
   const getL2Token = useL2Token();
-  const maxAmount = useMaxAmount();
+  const maxDeposit = useMaxDeposit();
 
   useEffect(() => {
     if (!selectedToken) {
@@ -44,27 +44,27 @@ export const Transfer = () => {
   useEffect(() => {
     if (selectedToken) {
       setHasInputError(false);
-      if (selectedToken.isLoading || Math.ceil(amount) === 0 || (isL1 && !maxAmount)) {
+      if (selectedToken.isLoading || Math.ceil(amount) === 0 || (isL1 && !maxDeposit)) {
         setIsButtonDisabled(true);
       } else {
         if (amount > selectedToken.balance) {
           setHasInputError(true);
           setErrorMsg(INSUFFICIENT_BALANCE_ERROR_MSG);
           setIsButtonDisabled(true);
-        } else if (isL1 && amount > maxAmount) {
+        } else if (isL1 && amount > maxDeposit) {
           setHasInputError(true);
-          setErrorMsg(MAX_AMOUNT_ERROR_MSG(maxAmount, symbol));
+          setErrorMsg(MAX_AMOUNT_ERROR_MSG(maxDeposit, symbol));
           setIsButtonDisabled(true);
         } else {
           setIsButtonDisabled(false);
         }
       }
     }
-  }, [amount, selectedToken, maxAmount, isL1]);
+  }, [amount, selectedToken, maxDeposit, isL1]);
 
   const onMaxClick = () => {
     try {
-      setAmount(String(Math.min(selectedToken.balance, Number(maxAmount))));
+      setAmount(String(Math.min(selectedToken.balance, Number(maxDeposit))));
     } catch (ex) {
       setAmount(selectedToken.balance);
     }
