+++ src/hooks/index.js
@@ -9,6 +9,8 @@ export * from './useWalletHandlerProvider';
 export * from './useFetchData';
 export * from './usePrevious';
 export * from './useTransfer';
+export * from './useTransferToL1';
+export * from './useTransferToL2';
 export * from './useConfig';
 export * from './useTransferProgress';
-export * from './useMaxAmount';
+export * from './useMaxDeposit';
